Extract tabs section to own component and update page headers
